feat: remove duplicates for search
