+++ package.json
@@ -37,7 +37,8 @@
     "react-dom": "^18.3.1",
     "siwe": "2.3.2",
     "viem": "^2.9.26",
-    "wagmi": "^2.5.21"
+    "wagmi": "^2.5.21",
+    "zod": "^3.23.6"
   },
   "devDependencies": {
     "@total-typescript/ts-reset": "^0.5.1",
