+++ src/services/index.ts
@@ -4,6 +4,7 @@ import { searchTalentProtocolUser } from "./talent-protocol";
 
 init(process.env.AIRSTACK_API_KEY!);
 
+
 export function generateRandomSequence(length: number) {
   let result = "";
   const characters =
@@ -15,110 +16,6 @@ export function generateRandomSequence(length: number) {
     counter += 1;
   }
   return result;
-}
-
-export async function searchSocialUser(querySearch: string) {
-  let addressToSearch = "";
-  let username = "";
-  if (querySearch.length === 42 && querySearch.startsWith("0x")) {
-    addressToSearch = querySearch;
-  } else {
-    username = querySearch;
-  }
-
-  // TODO: [possible improvement] add a 5 mins cache
-
-  const query = `query QueryUserOnLensAndFarcaster {
-        Socials(
-            input: {
-                filter: {
-                    dappName: { _in: [farcaster, lens] },
-                    identity: { ${addressToSearch.length > 0 ? `_eq: "${addressToSearch}"` : `_in: ["lens/@${username}","fc_fname:${username}", "${username}.eth"]`} }
-                },
-                blockchain: ethereum
-            }
-        ) {
-            Social {
-                userAddress
-                dappName
-                profileName
-                profileImage
-                userAssociatedAddresses
-                profileTokenId
-            }
-        }
-    }`;
-  const [
-    { data: talentProtocolData, error: talentProtocolError },
-    { data: airstackData, error: airstackError },
-  ] = await Promise.all([
-    searchTalentProtocolUser(querySearch),
-    fetchQuery(query),
-  ]);
-
-  if (airstackError) {
-    throw new Error(airstackError);
-  }
-  if (talentProtocolError && talentProtocolData !== null) {
-    throw new Error(talentProtocolError);
-  }
-
-  const filterFarcasterAddress = (
-    userAddress: string,
-    userAssociatedAddresses: string[],
-  ) => {
-    if (userAssociatedAddresses.length === 1) {
-      return userAddress;
-    }
-    const f = userAssociatedAddresses.filter((ua) => ua !== userAddress);
-    return f.length === 1 ? f[0] : userAddress;
-  };
-
-  // format results
-  const result: {
-    address: string;
-    username: string;
-    profile_image: string;
-    dapp: string;
-    profileTokenId: number;
-  }[] = [];
-  return result
-    .concat(
-      airstackData.Socials.Social
-        ? airstackData.Socials.Social.map(
-            (s: {
-              userAddress: string;
-              profileName: string;
-              dappName: string;
-              profileImage: string;
-              userAssociatedAddresses: string[];
-              profileTokenId: string;
-            }) => ({
-              address: filterFarcasterAddress(
-                s.userAddress,
-                s.userAssociatedAddresses,
-              ),
-              username: s.profileName,
-              profile_image: s.profileImage,
-              dapp: s.dappName,
-              profileTokenId: parseInt(s.profileTokenId, 10),
-            }),
-          )
-        : [],
-    )
-    .concat(
-      talentProtocolData
-        ? talentProtocolData.map((t) => ({
-            address: "",
-            username: t.user ? t.user.username : t.passport_profile!.name,
-            profile_image: t.user
-              ? t.user.profile_picture_url
-              : t.passport_profile!.image_url,
-            dapp: "talent-protocol",
-            profileTokenId: 0,
-          }))
-        : [],
-    );
 }
 
 export async function getNominationsFromFarcaster() {
